refactor: move modules into lorebook package (core/ + ui/)
Restructure the flat root-level modules into a proper Python package to
avoid namespace collisions from generic module names (matching.py,
features.py, image_utils.py) and enable `pip install -e .`.

- Move game-agnostic logic to lorebook/core/ (game_types, image_utils,
  matching, features, card_database, csv_manager).
- Move PySide6 widgets to lorebook/ui/ (styles, settings_window,
  main_window).
- Rewrite internal imports to absolute package paths.
- UI.py launcher and PhotoMatching.py shim/CLI stay at root; the shim
  re-exports from lorebook.core.* so existing imports (and tests) are
  unchanged.
- Add pyproject.toml (setuptools, requires-python >=3.11).
- Update CLAUDE.md architecture section and stale .json cache references.

All 47 unit tests pass; CLI and package imports verified.
