+++ UI.py
@@ -1,13 +1,13 @@
 # UI.py — Application entry point.
 #
-# All UI logic lives in:
+# All UI logic lives in lorebook/ui/:
 #   main_window.py    — MainWindow (camera, matching, CSV export)
 #   settings_window.py — SettingsWindow dialog
 
 from PySide6.QtCore import QTimer
 from PySide6.QtWidgets import QApplication
 
-from main_window import MainWindow, setup_logging
+from lorebook.ui.main_window import MainWindow, setup_logging
 
 setup_logging()
