+++ lorebook/__init__.py
@@ -1,0 +1,6 @@
+"""Lore-Book — desktop trading-card scanner package.
+
+Subpackages:
+    lorebook.core — game-agnostic logic (matching, features, database, CSV).
+    lorebook.ui   — PySide6 widgets (MainWindow, SettingsWindow, styles).
+"""
