+++ lorebook/core/__init__.py
@@ -1,0 +1,6 @@
+"""Core logic for Lore-Book: game types, image utils, matching, features,
+feature-database cache, and CSV management.
+
+Modules are imported individually (e.g. ``from lorebook.core.matching import
+find_best_matches``) so importing this package never eagerly loads TensorFlow.
+"""
