+++ lorebook/core/image_utils.py
