+++ lorebook/ui/__init__.py
@@ -1,0 +1,1 @@
+"""PySide6 UI for Lore-Book: MainWindow, SettingsWindow, and the app stylesheet."""
